chore: fixed minified error of layout overlay
